Re-organized code into packages
